Api concluida.

+++ pedido/src/main/java/br/com/compass/mercado/service/PedidoService.java
@@ -2,10 +2,14 @@ package br.com.compass.mercado.service;
 
 import br.com.compass.mercado.dto.request.RequestItemDto;
 import br.com.compass.mercado.dto.request.RequestPedidoDto;
+import br.com.compass.mercado.dto.response.ResponseMenssageDto;
 import br.com.compass.mercado.dto.response.ResponsePedidoDto;
 import br.com.compass.mercado.enums.Status;
+import br.com.compass.mercado.enums.StatusPagamento;
+import br.com.compass.mercado.exceptions.PedidoIndelibleException;
 import br.com.compass.mercado.exceptions.PedidoNotFoundException;
 import br.com.compass.mercado.exceptions.ValorDeDescontoInvalidoException;
+import br.com.compass.mercado.interfaces.ConsumerService;
 import br.com.compass.mercado.model.Pedido;
 import br.com.compass.mercado.repository.PedidoRepository;
 import br.com.compass.mercado.util.ValidaData;
@@ -18,7 +22,7 @@ import org.springframework.stereotype.Service;
 import java.math.BigDecimal;
 
 @Service
-public class PedidoService {
+public class PedidoService implements ConsumerService {
 
     @Autowired
     private PedidoRepository repository;
@@ -68,6 +72,12 @@ public class PedidoService {
 
     public void delete(Long id) {
         Pedido pedidoEntity = repository.findById(id).orElseThrow(PedidoNotFoundException::new);
+
+        if (pedidoEntity.getStatusPagamento().equals(StatusPagamento.APROVADO)
+                || pedidoEntity.getStatusPagamento().equals(StatusPagamento.REJEITADO) ) {
+            throw new PedidoIndelibleException();
+        }
+
         repository.delete(pedidoEntity);
     }
 
@@ -78,5 +88,20 @@ public class PedidoService {
 
         data.verificaData(dataDeCriacaoOferta, dataDeValidadeOferta);
         data.verificaValidade(dataDeValidadeOferta);
+    }
+
+    @Override
+    public void action(ResponseMenssageDto message) throws Exception {
+        Pedido pedido = repository.findById(message.getPedidoId()).orElseThrow(PedidoNotFoundException::new);
+
+        if (message.getStatus().equals("APPROVED")) {
+            pedido.setStatus(Status.FINALIZADO);
+            pedido.setStatusPagamento(StatusPagamento.APROVADO);
+        } else {
+            pedido.setStatus(Status.CANCELADO);
+            pedido.setStatusPagamento(StatusPagamento.REJEITADO);
+        }
+
+        repository.save(pedido);
     }
 }
